refactory - 코드정리
