+++ Assets/Scripts/Tile/Definitions/HexMetrics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f1cb9fcf2659b643b709ba2902b59a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
