+++ Assets/Scripts/Tile/Definitions/TileDefinitions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 484de497c326da04bbc61d23ed8964bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
