--- Assets/Scripts/Tile/HexCoordinates.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f2fccfced5788f45a65619cab18ce2b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
