+++ Assets/Scripts/Tile/Tile.cs
@@ -7,47 +7,24 @@ namespace Tile
     public class Tile : MonoBehaviour
     {
         public Color color;
-        public HexCoordinates coordinates;
-
-        public static readonly float SIZE_X = 1.73f;
-        public static readonly float SIZE_Y = 1.50f;
 
         private static GameObject[] tileResources = new GameObject[(int)TerrainType.NumTerrainType];
 
-        [SerializeField] private Transform renderObject;
-        [SerializeField] private GameObject debugGameObject;
-        [SerializeField] private TextMesh debugIndexText;
         [SerializeField] private Transform resourceRoot;
         [SerializeField] private bool isMovable = false;
         [SerializeField] private int moveCost = 1;
 
+        public HexCoordinates Coordinates { private set; get; }
+        public IndexPair IndexPair { private set; get; }
         public bool IsMovable { get { return isMovable; } }
         public int MoveCost { get { return moveCost; } }
-
         public int ContinentInfluence { get; set; }
-
-        public IndexPair IndexPair { private set; get; }
 
         public void Setup(IndexPair index)
         {
             IndexPair = index;
-
-#if UNITY_EDITOR
-            DebugTextToIndex();
-#endif
+            Coordinates = HexCoordinates.FromOffsetCoordinates(index.X, index.Y);
         }
-
-#if UNITY_EDITOR
-        public void CustomDebugText(string text)
-        {
-            debugIndexText.text = text;
-        }
-
-        public void DebugTextToIndex()
-        {
-            debugIndexText.text = $"({IndexPair.x}/{IndexPair.y})";
-        }
-#endif
 
         public void setupType(TerrainType terrainType, int featureType)
         {
@@ -65,8 +42,6 @@ namespace Tile
             var tileResourceTransform = Instantiate(tileResources[terrainTypeIndex]).transform;
             tileResourceTransform.parent = resourceRoot;
             tileResourceTransform.localPosition = Vector3.zero;
-            tileResourceTransform.tag = tag;
-            tileResourceTransform.gameObject.AddComponent<MeshCollider>();
         }
 
         private void setupMoveCost(TerrainType terrainType)
@@ -94,8 +69,6 @@ namespace Tile
             var tileResourceTransform = Instantiate(tileResource).transform;
             tileResourceTransform.parent = resourceRoot;
             tileResourceTransform.localPosition = Vector3.zero;
-            tileResourceTransform.tag = tag;
-            tileResourceTransform.gameObject.AddComponent<MeshCollider>();
         }
 #endif
     }
