+++ Assets/Scripts/Tile/TileCreaterEditor.cs
@@ -17,12 +17,11 @@ namespace Tile
         {
             "tilePrefab",
             "tileRange",
-            "hexGrid",
+            "tilesRoot",
             "hexMesh",
             "gridCanvas",
             "cellLabelPrefab",
             "defaultColor",
-            "touchedColor",
             "numOfMaxContinentTiles",
             "numOfLeastContinentTiles",
             "influenceOfContinent",
