+++ Assets/Scripts/Tile/TileTestEditor.cs
@@ -33,42 +33,11 @@ namespace Tile
 
         public override void OnInspectorGUI()
         {
-            if (GUILayout.Button("clear All Tile Debug Text")) TileHelper.ClearAllDebugText();
-            if (GUILayout.Button("show tile continent influence"))
-            {
-                TileModel.tiles.Where(x => x.ContinentInfluence != 0).All(x =>
-                {
-                    x.CustomDebugText(x.ContinentInfluence.ToString());
-                    return true;
-                });
-            }
-            if (GUILayout.Button("show tile index"))
-            {
-                TileModel.tiles.ForEach(x => x.CustomDebugText($"{x.IndexPair.x}/{x.IndexPair.y}"));
-            }
             if (GUILayout.Button("path finding"))
             {
-                TileHelper.ClearAllDebugText();
-
                 var startTile = TileHelper.GetTile(tileTester.findPathStart);
                 var endTile = TileHelper.GetTile(tileTester.findPathEnd);
                 tileTester.path = PathFinding.FindPath(startTile, endTile);
-
-                int count = 0;
-                Tile prevTile = null;
-                foreach (var tile in tileTester.path)
-                {
-                    if (prevTile)
-                    {
-                        tile.CustomDebugText($"{count += prevTile.MoveCost}");
-                    }
-                    else
-                    {
-                        tile.CustomDebugText($"{count}");
-                    }
-
-                    prevTile = tile;
-                }
             }
 
             serializedObject.Update();
