+++ Assets/Scripts/Tile/TileTester.cs
@@ -12,44 +12,5 @@ namespace Tile
         [SerializeField] public IndexPair findPathStart;
         [SerializeField] public IndexPair findPathEnd;
         [SerializeField] public List<Tile> path;
-
-        private List<Tile> cacheTiles = new List<Tile>();
-
-        private void Start()
-        {
-            // TODO : 타일 클릭 테스트
-            //var clickStream = Observable.EveryUpdate()
-            //.Where(_ => Input.GetMouseButtonDown(0))
-            //.Subscribe(_ =>
-            //{
-            //    var clickedObject = InputHelper.GetGameObjectFromScreenPointToRay(Input.mousePosition);
-
-            //    if (clickedObject) {
-            //        cameraModel.SetTarget(clickedObject.transform);
-            //    }
-            //});
-
-            var showNearTile = Observable.EveryUpdate()
-            .Where(_ => Input.GetKeyDown(KeyCode.F))
-            .Subscribe(_ =>
-            {
-                foreach (var cacheTile in cacheTiles)   
-                {
-                    cacheTile.DebugTextToIndex();
-                }
-                cacheTiles.Clear();
-
-                var tile = TileHelper.GetTile(getIndex);
-                tile.CustomDebugText("target");
-                cacheTiles.Add(tile);
-
-                foreach (var nearTile in TileHelper.GetNearTiles(tile))
-                {
-                    Debug.Log($"{nearTile.IndexPair.x} / {nearTile.IndexPair.y})");
-                    nearTile.CustomDebugText("near tile");
-                    cacheTiles.Add(nearTile);
-                }
-            });
-        }
     }
 }
